Deep-Learning/Fine-Tuning-Sentence-Embeddings/requirements.txt
torch>=2.1
sentence-transformers>=2.6
datasets
faiss-gpu
openai
python-dotenv
tqdm